Test if symlink exists first

+++ scap/checks/linkconf.sh
@@ -6,6 +6,10 @@ set -o pipefail
 DEPLOY_DIR=/srv/deployment/netbox/deploy
 
 # Create symlinks between configuration files and location netbox is expecting them
+if [ ! -f ${DEPLOY_DIR}/netbox/netbox/netbox/configuration.py ]; then
+	ln -s /etc/netbox-configuration.py ${DEPLOY_DIR}/netbox/netbox/netbox/configuration.py
+fi
 
-ln -s /etc/netbox-configuration.py ${DEPLOY_DIR}/netbox/netbox/netbox/configuration.py
-ln -s /etc/netbox-ldap.py ${DEPLOY_DIR}/netbox/netbox/netbox/ldap_config.py.py
+if [ ! -f ${DEPLOY_DIR}/netbox/netbox/netbox/ldap_config.py.py ]; then
+    ln -s /etc/netbox-ldap.py ${DEPLOY_DIR}/netbox/netbox/netbox/ldap_config.py.py
+fi
